[BE] Move result saving logic to another file
